+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Added fallback for Mailfence and other non-quote-compliant providers - [#209](https://github.com/hkdb/aerion/issues/209)
 - Fixed CardDav remove provider code path to not leave orphaned contacts in db
 - Fixed Sent/Draft folder message listing - [#227](https://github.com/hkdb/aerion/issues/227)
+- Fixed unified inbox actions - [#234](https://github.com/hkdb/aerion/issues/234)
 
 
 **v0.2.5 - 05-27-2026**